refactor(lib): Split up anubis.go into some smaller files.  (#379)
* refactor(logging): centralize logger creation in GetLogger function


* refactor(logging): rename GetLogger to GetRequestLogger for clarity


* refactor: streamline error handling and response methods


* refactor(lib): Split anubis.go up into some smaller specialized methods


* refactor(http): simplify error response handling by using respondWithStatus


* chore(lib): run goimports


---------
